+++ packages/capabilities/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llm-ports/capabilities
 
+## 0.1.0-alpha.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @llm-ports/core@0.1.0-alpha.31
+
 ## 0.1.0-alpha.30
 
 ### Patch Changes